Introducing build flags to avoid including unessential functionality on android devices.
